adaptações para o plano de aula
