+++ src/br/com/alura/kotlin/collection/map/Aluno.kt
@@ -1,0 +1,7 @@
+package br.com.alura.kotlin.collection.map
+
+data class Aluno (
+    val rm: String,
+    val nome: String,
+    val sobreNome: String
+)
