+++ src/br/com/alura/kotlin/collection/map/Usuario.kt
@@ -1,7 +1,7 @@
 package br.com.alura.kotlin.collection.map
 
 data class Usuario (
-        val nome: String,
-        val sobreNome: String,
-        val email: String
+    val nome: String,
+    val sobreNome: String,
+    val email: String
 )
